feat(kit): moved extractAffixes and fixed issues
